scripts/genome_download/pyproject.toml
[tool.poetry]
name = "production"
version = "0.1.0"
description = "Rfam production code"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
click = "^8.1.3"
biopython = "^1.79"
Cython = "^0.29.30"
sqlitedict = "^2.0.0"
attrs = "^21.4.0"
cattrs = "^22.1.0"
flit-scm = "^1.7.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
